* "Routes" installed * Updated App.jsx * Costo Sanitario moved to "pages" * Updated Header.
