+++ package.json
@@ -18,6 +18,7 @@
     "react-chartjs-2": "^5.3.0",
     "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
+    "react-router-dom": "^7.6.1",
     "tailwindcss": "^4.1.7"
   },
   "devDependencies": {
