+++ src/pages/Fertilizacion.jsx
@@ -1,0 +1,5 @@
+function Fertilizacion() {
+  return <div>Fertilizacion</div>;
+}
+
+export default Fertilizacion;
